Upload example sql and file source content
In order to install the examples, run the provided sql script. It will create the data sources as well as virtual schemas, views and procedures from the masterclass track.
The zip file is the content for the file source - please point the DV file type data source to the folder you extracted it to.